Add error handling for Arango Server Password update migration

+++ scripts/update-arango-password.sh
@@ -15,7 +15,14 @@ echo New arango password stored in ${FOLDERDIR}/arango.txt file
 #echo Generated new arango password: $new_arango_password
 
 touch arango-password-script.js
-echo 'require("@arangodb/users").replace("root", ARGUMENTS[0]);' > arango-password-script.js
+
+echo 'try {'                                                        > arango-password-script.js
+echo '    require("@arangodb/users").replace("root", ARGUMENTS[0]);'> arango-password-script.js
+echo '    print("SUCCESS");'                                        > arango-password-script.js
+echo '} catch (error) {'                                            > arango-password-script.js
+echo '    print("FAILURE");'                                        > arango-password-script.js
+echo '    print(error);'                                            > arango-password-script.js
+echo '}'                                                            > arango-password-script.js
 
 echo Updating arango server password
 
@@ -25,11 +32,17 @@ supervisorctl start arango
 
 sleep 10s
 
-/usr/bin/arangosh --server.password "" --javascript.execute arango-password-script.js ${new_arango_password}
+status=$(/usr/bin/arangosh --server.password "" --javascript.execute arango-password-script.js ${new_arango_password})
 
 supervisorctl stop arango
 sed -i 's/authentication = false/authentication = true/g' /etc/arangodb3/arangod.conf
 supervisorctl start arango
+
+if [[ $status != "SUCCESS" ]];
+then
+    mv ${FOLDERDIR}/arango.txt ${FOLDERDIR}/arango_failed.txt
+    exit 1
+fi
 
 rm arango-password-script.js
